Preserve order of event sources
When we filter out duplicates, we keep the one that we first saw. This
is problematic since we don't know the order in which events come in
from the different sources. By moving the construction of the `allItems`
array to after all the promises have resolved, we will maintain the
order between different sources.

Also moved the filtering out duplicates to above the sort() so that we
can make certain events consistently "stick".